pyproject.toml
[project]
name = "bookie-evaluator"
version = "0.1.0"
description = "Edge finding in betting"
authors = [
    {name = "namatanda",email = "mwagichimu@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.31.0)",
    "pandas (>=2.1.1)",
    "numpy (>=1.26.0)",
    "matplotlib (>=3.8.0)",
    "scikit-learn (>=1.3.0)",
    "python-dotenv (>=1.0.0)",
    "tqdm (>=4.66.1)",
    "tabulate (>=0.9.0)",
    "colorama (>=0.4.6)",
    "xgboost (>=2.0.0)",
    "lightgbm (>=4.1.0)",
    "shap (>=0.43.0)",
    "joblib (>=1.3.2)",
    "optuna (>=3.3.0)",
    "seaborn (>=0.13.2,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
